Разбиение проекта на модули
