+++ js/database.js
@@ -1,0 +1,50 @@
+const NAMES = [
+  'Алексей',
+  'Виктория',
+  'Дмитрий',
+  'Екатерина',
+  'Иван',
+  'Мария',
+  'Павел',
+  'София',
+  'Тимофей',
+  'Юлия'];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const DESCRIPTIONS = [
+  'Вид из окна моего отеля!',
+  'Идем на пляж.',
+  'Нашли укромное место с видом на пляж',
+  'Сегодня загараем и делаем фотки!',
+  'Подача блюд в местном ресторане)',
+  'Увидел во дворе такую красавицу!',
+  'Сажусь на диету!',
+  'Наготовили морса!',
+  'А у нас сегодня была экстренная посадка на пляже.',
+  'Нашла куртой органайзер для обуви на маркетплейсе.',
+  'Дорога на пляж.',
+  'Купил новую машину!',
+  'Легкий перекус на Бали',
+  'У нашего Барсика фотосессия',
+  'Классные тапки!',
+  'Сделал фото пока летел.',
+  'Наконец-то побывал на концерте.',
+  'Был сегодня на выставке ретро автомобилей.',
+  'О! Мне такие надо!',
+  'Внутрений двор нашего отеля.',
+  'Сегодня буду готовить такую вкусняху!',
+  'Снял красивый закат.',
+  'Сфоткала вот такого красавца недалеко от пляжа.',
+  'Ура! Наконец-то побывал на их концерте!',
+  'Побывали на сафари!',
+];
+
+export { NAMES, MESSAGES, DESCRIPTIONS };
